backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pandas==2.1.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
tensorflow-cpu==2.14.0
pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.0
supabase==2.4.1
web3==6.11.3
